# perf: serialize into a string, then write to disk once