Another performance improvement. It turns out that the 2.4.0 changes were actually slower for small to medium objects compared to the simple values we tested.

```
2.3.0
10x, 0 keys, 100000 changes
  4160.50 ms / attempt
  2403.56 changes/s

2.4.0
10x, 0 keys, 100000 changes
  5698.30 ms / attempt
  1754.91 changes/s

2.4.0, maxBufferedCommands = 1000
10x, 0 keys, 100000 changes
  5669.80 ms / attempt
  1763.73 changes/s

2.4.0, sizeFactor = 5
10x, 0 keys, 100000 changes
  5686.40 ms / attempt
  1758.58 changes/s
```

This change bumps it to:
```
2.4.0, write once
10x, 0 keys, 100000 changes
  485.70 ms / attempt
  20588.84 changes/s
```